Move stock organization logic from Facility
Migrate all logic to use bounded context schemas
