+++ lib/open_pantry/food_selection/food_selection.ex
@@ -7,10 +7,49 @@ defmodule OpenPantry.FoodSelection do
   alias OpenPantry.Repo
   alias Ecto.Multi
   alias OpenPantry.StockDistribution
-  alias OpenPantry.Stock
+  alias OpenPantry.FoodSelection.Stock
+  alias OpenPantry.FoodSelection.CreditType
   alias OpenPantry.UserCredit
   alias OpenPantry.UserOrder
-  alias OpenPantry.CreditType
+  alias OpenPantry.Facility
+
+  @spec stock_by_type(Facility.t) :: list(tuple())
+  def stock_by_type(facility = %Facility{id: id}) do # way more data than needed, but one query! :-/
+    now = DateTime.utc_now
+    from(credit_type in CreditType,
+    join: stocks in assoc(credit_type, :stocks),
+    where: stocks.arrival < ^now,
+    where: stocks.expiration > ^now,
+    where: ^id == stocks.facility_id,
+    order_by: credit_type.inserted_at,
+    preload: [stocks: [food: :food_group]])
+    |> Repo.all
+    |> Enum.map(&({&1.name, &1.id, &1.stocks}))
+    |> Enum.uniq
+    |> append_meals_if_any(facility)
+  end
+
+  @spec append_meals_if_any(list(tuple()), Facility.t) :: list(tuple())
+  def append_meals_if_any(food_stocks, facility) do
+    meal_stocks = meal_stocks(facility)
+    food_stocks ++  if Enum.any?(meal_stocks) do
+                      [{"Meals", :meals, meal_stocks}]
+                    else
+                      []
+                    end
+  end
+
+  @spec meal_stocks(Facility.t) :: list(Stock.t)
+  def meal_stocks(%Facility{id: id}) do
+    now = DateTime.utc_now
+    from(stocks in Stock,
+    where: stocks.arrival < ^now,
+    where: stocks.expiration > ^now,
+    where: ^id == stocks.facility_id,
+    where: fragment("? IS NOT NULL", stocks.meal_id),
+    preload: [:meal])
+    |> Repo.all
+  end
 
   @spec adjust_stock(integer(), integer(), integer(), integer() ) :: StockDistribution.t
   def adjust_stock(stock_id, type_id, quantity, user_id) do
