+++ lib/open_pantry/web/models/facility.ex
@@ -31,42 +31,4 @@ defmodule OpenPantry.Facility do
     |> validate_required([:name])
   end
 
-  @spec stock_by_type(Facility.t) :: list(tuple())
-  def stock_by_type(facility = %Facility{id: id}) do # way more data than needed, but one query! :-/
-    now = DateTime.utc_now
-    from(credit_type in CreditType,
-    join: stocks in assoc(credit_type, :stocks),
-    where: stocks.arrival < ^now,
-    where: stocks.expiration > ^now,
-    where: ^id == stocks.facility_id,
-    order_by: credit_type.inserted_at,
-    preload: [stocks: [food: :food_group]])
-    |> Repo.all
-    |> Enum.map(&({&1.name, &1.id, &1.stocks}))
-    |> Enum.uniq
-    |> append_meals_if_any(facility)
-  end
-
-  @spec append_meals_if_any(list(tuple()), Facility.t) :: list(tuple())
-  def append_meals_if_any(food_stocks, facility) do
-    meal_stocks = meal_stocks(facility)
-    food_stocks ++  if Enum.any?(meal_stocks) do
-                      [{"Meals", :meals, meal_stocks}]
-                    else
-                      []
-                    end
-  end
-
-  @spec meal_stocks(Facility.t) :: list(Stock.t)
-  def meal_stocks(%Facility{id: id}) do
-    now = DateTime.utc_now
-    from(stocks in Stock,
-    where: stocks.arrival < ^now,
-    where: stocks.expiration > ^now,
-    where: ^id == stocks.facility_id,
-    where: fragment("? IS NOT NULL", stocks.meal_id),
-    preload: [:meal])
-    |> Repo.all
-  end
-
 end
