+++ lib/open_pantry/web/models/stock.ex
@@ -2,6 +2,7 @@ defmodule OpenPantry.Stock do
   use OpenPantry.Web, :model
   alias OpenPantry.Repo
   alias OpenPantry.Food
+  alias OpenPantry.FoodSelection
   alias OpenPantry.Meal
   alias OpenPantry.Offer
   alias OpenPantry.Facility
